hotfix: printer

+++ src/modules/io/printer.h
@@ -20,7 +20,7 @@ typedef struct __printer {
   print_list *head;
   print_list *tail;
   char printer_buff[PRINTER_WINDOW][MAX_PRINTER_OUTPUT];
-  u8 buff_last;
+  i16 buff_last;
   bool active;
 } printer;
